requirements.txt
# Templonix Lite - Streamlined Dependencies

# Core AI & LLM
pydantic==2.11.2
pydantic-settings==2.8.1

# MCP Integration
mcp==1.14.1
starlette>=0.27
uvicorn>=0.31.1

# Database & Memory - Now using FAISS locally
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# Email & Calendar (optional)
google-api-python-client==2.152.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# Web & HTTP
aiohttp==3.11.16
httpx==0.28.1
requests==2.32.3

# File & Document Processing (for expansions)
pillow==11.1.0
pypdf==4.0.0
python-docx=1.1.0

# Utilities
python-dotenv==1.1.0
PyYAML==6.0.2
click==8.1.8
rich==13.9.4

# Development & Testing
pytest==8.3.5
black==25.1.0
isort==6.0.1
flake8==7.3.0
mypy==1.15.0
pytest-cov==6.0.0
pytest-asyncio==0.25.2
pytest-mock==3.15.0

# Common Dependencies
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
urllib3==2.3.0
packaging==24.2
typing-extensions==4.13.0
anyio==4.9.0
sniffio==1.3.1